adjusted product fetch and create

+++ database/migrations/2022_10_04_144106_create_products_variation_table.php
@@ -18,7 +18,7 @@ class CreateProductsVariationTable extends Migration
             $table->bigInteger('products_id');
             $table->string('name');
             $table->string('detail');
-            $table->integer('products_price');
+            $table->double('products_price');
             $table->softDeletes();
             $table->timestamps();
         });
